chore: Move `changeAccount` to own file
