+++ source/scripts/Background/controllers/message-handler/requests.ts
@@ -1,8 +1,7 @@
-import { browser } from 'webextension-polyfill-ts';
-
 import { EthProvider } from 'scripts/Provider/EthProvider';
 import { SysProvider } from 'scripts/Provider/SysProvider';
 
+import { changeAccount } from './change-account';
 import { Message } from './types';
 
 /**
@@ -47,46 +46,4 @@ export const methodRequest = async (
   if (!method) throw new Error('Unknown method');
 
   return await method(message.data.args);
-};
-
-/**
- * Opens a popup to select another account to connect
- *
- * @return `true` if the selected account was changed
- */
-const changeAccount = async (
-  network: string,
-  origin: string,
-  isPendingWindow: () => boolean,
-  setPendingWindow: (isPending: boolean) => void
-) => {
-  const { dapp, createPopup } = window.controller;
-
-  const isConnected = dapp.isConnected(origin);
-
-  if (isPendingWindow() || !isConnected) return;
-
-  const popup = await createPopup('change-account', { network, origin });
-
-  setPendingWindow(true);
-
-  return new Promise<boolean>((resolve) => {
-    window.addEventListener(
-      'accountChange',
-      (event: CustomEvent) => {
-        if (event.detail.origin === origin) {
-          setPendingWindow(false);
-          resolve(true);
-        }
-      },
-      { once: true, passive: true }
-    );
-
-    browser.windows.onRemoved.addListener((id) => {
-      if (id === popup.id) {
-        setPendingWindow(false);
-        resolve(false);
-      }
-    });
-  });
 };
